Refactor list item renderers
